Add documentation in README

+++ real-time-scheduler/real_time_queueing.h
@@ -17,7 +17,6 @@
 class Packet {  // packet with birth time and deadline
 public:  // NOLINT
     Packet(int i = -1, int j = -1) {birth_time_ = i; deadline_ = j;}
-        // TODO(Veggente): initializer list
     int birth_time() const {return birth_time_;}
     int deadline() const {return deadline_;}
     int delay_bound() const {return deadline_-birth_time_+1;}
@@ -28,7 +27,6 @@ private:  // NOLINT
 
 class QueueingSystem {
 public:  // NOLINT
-         // TODO(Veggente): unnecessary methods cleanup
     QueueingSystem(const BooleanMatrix &m,
                    Policy s, Ratios q, int b,
                    int d, const std::string &f, int n);
@@ -52,11 +50,11 @@ public:  // NOLINT
     int quarter_point();  // quarter point integer among num_iterations
     int half_point();  // half point integer among num_iterations
     void update_stability_counter();
-    double lower_deficit_sum() const {return lower_deficit_sum_;}
-    double upper_deficit_sum() const {return upper_deficit_sum_;}
+    int64_t lower_deficit_sum() const {return lower_deficit_sum_;}
+    int64_t upper_deficit_sum() const {return upper_deficit_sum_;}
     double stability_ratio();
-    long sum_cumulative_arrival();
-    long sum_cumulative_throughput();
+    int64_t sum_cumulative_arrival();
+    int64_t sum_cumulative_throughput();
 private:  // NOLINT
     Queues per_link_queue_;
     Counters per_link_deficit_;
@@ -69,10 +67,8 @@ private:  // NOLINT
     int system_clock_;
     int max_delay_bound_;
     std::string output_filename_;
-    double lower_deficit_sum_;  // quarter-to-half deficit sum
-                                // TODO(Veggente): use long instead of double
-    double upper_deficit_sum_;  // half-to-whole deficit sum
-                                // TODO(Veggente): use long instead of double
+    int64_t lower_deficit_sum_;  // quarter-to-half deficit sum
+    int64_t upper_deficit_sum_;  // half-to-whole deficit sum
     int num_iterations_;
     Counters per_link_cumulative_throughput_;
     Counters per_link_cumulative_arrival_;
